+++ src/settings/InventorySettings/InventorySettings.js
@@ -45,6 +45,7 @@ import ClassificationBrowseSettings from '../ClassificationBrowseSettings';
 import SubjectSourcesSettings from '../SubjectSourcesSettings';
 import SubjectTypesSettings from '../SubjectTypesSettings';
 import DisplaySettings from '../DisplaySettings';
+import CardsPerVersionHistoryPage from '../CardsPerVersionHistoryPage';
 import CallNumberBrowseSettings from '../CallNumberBrowseSettings';
 import {
   flattenCentralTenantPermissions,
@@ -93,6 +94,17 @@ const InventorySettings = (props) => {
             route: 'displaySettings',
             label: <FormattedMessage id="ui-inventory.settings.section.displaySettings" />,
             component: DisplaySettings,
+            perm: 'ui-inventory.settings.displaySettings',
+          },
+        ],
+      },
+      {
+        label: <FormattedMessage id="ui-inventory.settings.heading.versionHistory" />,
+        pages: [
+          {
+            route: 'cardsPerPage',
+            label: <FormattedMessage id="ui-inventory.settings.section.cardsPerPage" />,
+            component: CardsPerVersionHistoryPage,
             perm: 'ui-inventory.settings.displaySettings',
           },
         ],